feat: update OrderRequest and Product entities, and implement OrderService for order creation

+++ practise/src/main/java/com/example/practise/entities/OrderItem.java
@@ -35,7 +35,7 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(name = "price_at_purchase", nullable = false)
-    private BigDecimal priceAtPurchase;
+    private Integer priceAtPurchase;
 
     // some relations
     @JsonBackReference
